app/requirements.txt
fastapi==0.104.1
sqlalchemy==2.0.23
uvicorn==0.24.0
bcrypt==4.1.1
pydantic==2.5.0
pydantic-settings==2.1.0
sqlmodel==0.0.14
psycopg[binary]==3.1.13
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
python-dotenv==1.0.0
pika==1.3.2
requests==2.32.4
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
api-analytics==1.2.7
# Document processing libraries
PyPDF2==3.0.1
python-docx==1.1.0
docx2txt==0.8